move remote config tool
